% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{CheckSignalRange}
\alias{CheckSignalRange}
\title{Check if all signals in a signal pack are in the right range}
\usage{
CheckSignalRange(signal.pack)
}
\arguments{
\item{signal.pack}{a list of signals, xts objects}
}
\description{
Check if all signals in a signal pack are in the right range
}
