% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{CleanSubset}
\alias{CleanSubset}
\title{Clean a list of dataframe with same available `clean` index}
\usage{
CleanSubset(var.list)
}
\arguments{
\item{var.list}{a list of dataframe/xts, each is for one asset}
}
\value{
a new list of dataframe/xts with the SAME index, no NA and no zero
}
\description{
Clean a list of dataframe with same available `clean` index
}
