% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{CutSeriesQuantile}
\alias{CutSeriesQuantile}
\title{Cut range for one single series using quantiles}
\usage{
CutSeriesQuantile(series, upper = 1 - 0.001, lower = 0.001)
}
\arguments{
\item{series}{xts object}

\item{upper, lower}{numeric, quantiles, i.e. 0.99, 0.01}
}
\description{
Cut range for one single series using quantiles
}
