% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{CutSeriesRange}
\alias{CutSeriesRange}
\title{Cut range for one single series}
\usage{
CutSeriesRange(series, upper, lower)
}
\arguments{
\item{series}{xts object}

\item{upper, lower}{numeric limits}
}
\description{
Cut range for one single series
}
