% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{DecompMetrics}
\alias{DecompMetrics}
\title{Decompose the huge metrics dataframe into a list}
\usage{
DecompMetrics(metrics)
}
\arguments{
\item{metrics}{dataframe from OutputWrapper}
}
\value{
a list of dataframe, each df for one variable
  the names of the list are variable names, in each df
  colnames are adjusted to be just names of methods,
  to save space
}
\description{
Decompose the huge metrics dataframe into a list
}
