% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{DEtoUD}
\alias{DEtoUD}
\title{Transfer DM and EM (original DRI) signal to US and DM ratio signal}
\usage{
DEtoUD(DE.signal, us.init = 0.6, dm.init = 6/7)
}
\arguments{
\item{DE.signal}{xts series of DM and EM columns (original DRI signal),
ranging from 0 to 1}

\item{us.init, dm.init}{numeric initial value of US and DM proportions}
}
\value{
list of 2 signals, US and DM, representing US's proportion in
  overall portfolio and DM's proportion in international (EE) portfolio
}
\description{
Transfer DM and EM (original DRI) signal to US and DM ratio signal
}
