% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{DistSQ}
\alias{DistSQ}
\title{Calculate Square Distance between ith and jth in Dens list}
\usage{
DistSQ(dens, i, j)
}
\arguments{
\item{dens}{density list generated from GenXXXX}

\item{i, j}{int}
}
\description{
Calculate Square Distance between ith and jth in Dens list
}
