% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoot.R
\name{DistUR}
\alias{DistUR}
\title{Calculate Utility Risk Distance between ith and jth in Dens list}
\usage{
DistUR(dens, i, j, idx, gamma = 0.6)
}
\arguments{
\item{dens}{density list generated from GenXXXX}

\item{i, j}{int}

\item{idx}{index of densities, can be generated again out of the funcion
as the index of density is not returned in GenXXXX function}

\item{gamma}{float parameter of CRRA, not very influential}
}
\description{
asymmetric adjusted weights for 2 tails, higher weights on left
}
