% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artillery.R
\name{FacToStr}
\alias{FacToStr}
\title{Convert factor columns to character string}
\usage{
FacToStr(raw.df, exclude = NULL)
}
\arguments{
\item{raw.df}{dataframe with factor columns (actual numbers)}

\item{exclude}{vector of strings/names of columns excluded
from the operation, default: NULL}
}
\value{
dataframe with factor columns transferred to numeric,
  for '#%' type columns, change to '#/100' strings, can be
  evaluated using eval() as next step
}
\description{
More conditions of converting factors to expressions
  strings can be added after the gsub() function
}
