% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{GenASH}
\alias{GenASH}
\title{Generate density list for groups using ASH}
\usage{
GenASH(groups, nbin = 128)
}
\arguments{
\item{groups}{a list of conditional variables from GenCondGroups}

\item{bw, gs}{hyperparameters}
}
\value{
list of density series
}
\description{
Generate density list for groups using ASH
}
