% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{GenASHBand}
\alias{GenASHBand}
\title{Generate Bootstrap Band for ASH}
\usage{
GenASHBand(groups, nbin = 128, n = 1000, pctg = 0.8)
}
\arguments{
\item{groups}{conditional variable series from GenCondGroups}

\item{nbin}{number of bin hyperparameter}

\item{n}{number of times for boostrapping}

\item{pctg}{float percentage of total sample for subsample}
}
\description{
Generate Bootstrap Band for ASH
}
