% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoot.R
\name{GenBinShDens}
\alias{GenBinShDens}
\title{Generate Binary Shrinked Groups' Densities}
\usage{
GenBinShDens(dens, shlist)
}
\arguments{
\item{dens}{list of densities of subgroups}

\item{shlist}{list of shrinked groups, from GenBinShHC}
}
\value{
list of densities in shrinked groups (default 2)
}
\description{
Generate Binary Shrinked Groups' Densities
}
