% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoot.R
\name{GenBinShHC}
\alias{GenBinShHC}
\title{Generate Binary Shrinked HC Clusters}
\usage{
GenBinShHC(merge)
}
\arguments{
\item{merge}{matrix from hclust, hc$merge}
}
\value{
a list of 2 vectors, showing subgroups in shrinked groups
}
\description{
Generate Binary Shrinked HC Clusters
}
