% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet-rollmat.R
\name{GenBivRet}
\alias{GenBivRet}
\title{Generate a list for bivariate returns}
\usage{
GenBivRet(ret)
}
\arguments{
\item{ret}{xts object of individual returns}
}
\value{
list of xts object, each is a pair of returns
  extracted from the input ind.returns
}
\description{
Generate a list for bivariate returns
}
