% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoot.R
\name{GenBKDE}
\alias{GenBKDE}
\title{Generate density list for groups using BKDE KernSmooth}
\usage{
GenBKDE(groups, bw = 0.003, gs = 128L, range = NULL)
}
\arguments{
\item{groups}{a list of conditional variables from GenCondGroups}

\item{bw, gs}{hyperparameters}
}
\value{
list of density series
}
\description{
Generate density list for groups using BKDE KernSmooth
}
