% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{GenBKDEBand}
\alias{GenBKDEBand}
\title{Generate Bootstrap Band for KernSmooth}
\usage{
GenBKDEBand(groups, bw = 0.003, gs = 128, n = 100, pctg = 0.8)
}
\arguments{
\item{groups}{conditional variable series from GenCondGroups}

\item{bw, gs}{hyperparams for KernSmoothing}

\item{n}{number of times for boostrapping}

\item{pctg}{float percentage of total sample for subsample}
}
\description{
Generate Bootstrap Band for KernSmooth
}
