% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoot.R
\name{GenBoolSignal}
\alias{GenBoolSignal}
\title{Generate bool signal vector for a given continuous 0-1 (quantile) signal}
\usage{
GenBoolSignal(signal, n.group = 10, cuts = seq(0, 1, 1/n.group))
}
\arguments{
\item{signal}{xts numerical from 0 to 1, i.e. from GenEmpQuantileVec}

\item{n.group}{int number of groups divided}

\item{cuts}{vector of number used to divide groups}
}
\description{
Generate bool signal vector for a given continuous 0-1 (quantile) signal
}
\examples{
  GenBoolSignal(quantiles, 6, cuts = c(0, 1/10, 3/10, 0.5, 1-3/10, 1-1/10, 1))
  GenBoolSignal(quantiles, 10)
}
