% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{GenCondGroups}
\alias{GenCondGroups}
\title{Generate conditional groups based on bool signals and a variable/return xts}
\usage{
GenCondGroups(ret, bools)
}
\arguments{
\item{ret}{xts variable/return}

\item{bools}{a 2d xts bool, Vector{1dxts{bool}}}
}
\value{
a list of 1dxts, same size as bools
}
\description{
Generate conditional groups based on bool signals and a variable/return xts
}
