% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf-metrics.R
\name{GenCoreMetrics}
\alias{GenCoreMetrics}
\title{Get core performance metrics for monthly log returns}
\usage{
GenCoreMetrics(total.ret, benchmark.ret = total.ret,
  period = c("Monthly", "Daily"), market.ret = NULL, rf = 0)
}
\arguments{
\item{total.ret}{xts object monthly log returns}

\item{market.ret}{xts same size and index of total.ret}

\item{rf}{numeric monthly risk free rate, default: 3\%}

\item{banchmark.ret}{xts same size and index of total.ret}

\item{peroid}{period of returns, for diff annualized factors}
}
\description{
Get core performance metrics for monthly log returns
}
