% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet.R
\name{GenEmpQuantileVec}
\alias{GenEmpQuantileVec}
\title{Generate Empirical Quantile for a vector with its own info}
\usage{
GenEmpQuantileVec(vec)
}
\arguments{
\item{vec}{vector/1dxts}
}
\value{
vector/1dxts of empirical quantile based on its emp distr
}
\description{
Generate Empirical Quantile for a vector with its own info
}
