% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoot.R
\name{GenHC}
\alias{GenHC}
\title{Generate HC results for Shooting a Bullet for a Target}
\usage{
GenHC(bullet, target)
}
\arguments{
\item{bullet, target}{xts time series object, might have diff period}
}
\value{
HC results
}
\description{
Generate HC results for Shooting a Bullet for a Target
}
