% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoot.R
\name{GenHeights}
\alias{GenHeights}
\title{Generate Top 3 Heights, shooting a bullet}
\usage{
GenHeights(bullet = asym1m, target = ret.xlk \%>\% lag.xts(-1),
  n.top = 3)
}
\arguments{
\item{bullet, target}{xts time series object, might have diff period}

\item{n.top}{int number of top heighs required for the report}
}
\value{
vector of top heights using the bullet hitting the target
}
\description{
Generate Top 3 Heights, shooting a bullet
}
