% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mgmt.R
\name{GenInitSet}
\alias{GenInitSet}
\title{Generate Initial Set for a certain universe}
\usage{
GenInitSet(tickers, columns, start.date, append)
}
\arguments{
\item{tickers}{vector of strings}

\item{columns}{vector of strings for variables, colnames}

\item{start.date}{date, i.e. as.Date("2013-01-01")}

\item{append}{a list of xts dataframe that will be added in the output
and in the CleanSubset function, its availability will have impact on
the final outputs}
}
\value{
a list of xts dataframes with identical index, including traded assets
  and other useful time series (not necessary an asset, i.e. VIX)
}
\description{
Generate Initial Set for a certain universe
}
