% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-ind.R
\name{GenMonthlyRet}
\alias{GenMonthlyRet}
\title{Function Wrapper for xts::apply.monthly}
\usage{
GenMonthlyRet(daily.returns)
}
\arguments{
\item{daily.returns}{xts object with individual returns}
}
\value{
xts object with monthly log return
}
\description{
the reason for this function is that xts:apply.monthly
  only take one column (or the first column if called with a df), and
  xts::to.monthly has outputs with multiple columns (OHLC)
}
