% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{GenOutputs}
\alias{GenOutputs}
\title{Generate complete outputs for a list of signals, weights, and trets}
\usage{
GenOutputs(weights, trets, returns, returns.period = c("Monthly",
  "Daily"), signal.names, bmk.ret = trets[[1]])
}
\arguments{
\item{weights}{a list of xts dataframes, representing weights time series, generated
from GenWeights, with names properly set as signal.names}

\item{trets}{a list of xts series, total returns of signals, generated from GenTrets,
also with names properly set as signal.names}

\item{returns}{a xts dataframe with columns being assets traded in the portfolio, should
have the same number of columns as member in weights}

\item{returns.period}{string, indicating periodicty of returns, can be used to calculate daily
performance metrics if passed a daily return dataframe and monthly weights}

\item{signal.names}{should be passed from outside environment, a vector of string, previously
set as unlist(lapply(signal.pack, function(x) colnames(x)[1])) being default}
}
\value{
a list of metrics dataframe and pnl xts object with proper colnames
}
\description{
Generate complete outputs for a list of signals, weights, and trets
}
