% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet-rollmat.R
\name{GenPairName}
\alias{GenPairName}
\title{Generate names for pair-wise matrix given individual names}
\usage{
GenPairName(names)
}
\value{
vector of string, names for vectorized pairwise matrix
  following row-first order and only for upper triangular
  matrix of the square vcov matrix
}
\description{
Generate names for pair-wise matrix given individual names
}
