% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-ind.R
\name{GenRelScore}
\alias{GenRelScore}
\title{Generate dynamic relative scores for input matrix}
\usage{
GenRelScore(data, data.period, rolling.period, burn.in.period = NULL,
  scoring.method, transfer.method = NULL, output.period = "monthly")
}
\arguments{
\item{data}{xts object vectorized named rolling (vcov) dataframe}

\item{data.period}{string 'monthly' or 'daily', should match the data}

\item{rolling.period}{numeric number of data.period should be included
in each calculation in the rolling window}

\item{burn.in.period}{numeric number of period (specified in data.period)
that's used to generate the first realized value, can be included
in the rolling period.}

\item{scoring.method}{function getting df and returning a vec}

\item{transfer.method}{function used to transfer daily result to monthly
when output.period != data.period, from a vec to a value}

\item{output.period}{string 'monthly' or 'daily', frequency of return
xts object, default: 'monthly'}
}
\value{
xts object with monthly
}
\description{
Generate dynamic relative scores for input matrix
}
