% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet.R
\name{GenSeqValue}
\alias{GenSeqValue}
\title{Generate sequential value for the "i, j" point in a n squared matrix}
\usage{
GenSeqValue(n, i, j)
}
\arguments{
\item{n, i, j}{int matrix is nth squared, point pos is i,j}
}
\value{
int seq value starting from upper left, row first, only for upper tri
}
\description{
Generate sequential value for the "i, j" point in a n squared matrix
}
