% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet.R
\name{GenSpike}
\alias{GenSpike}
\title{Generate scaled z scores for volatility spikes}
\usage{
GenSpike(df)
}
\arguments{
\item{df}{xts object of rolling vcov matrix, monthly or daily}
}
\value{
numeric vector of (scaled) spike values
}
\description{
Generate scaled z scores for volatility spikes
}
