% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{GenSubidxSD}
\alias{GenSubidxSD}
\title{Generate sub period index based on robust rolling std range (filtering/regime id)}
\usage{
GenSubidxSD(series, winlen = 36, sd.threshold = 3, smooth.winlen = 3,
  IF.LAG = FALSE)
}
\arguments{
\item{series}{xts object, benchmark time series, monthly for now}

\item{winlen}{numeric, rolling window length for calculating sd and mean}

\item{sd.threshold}{integer, used for the boundary of standard deviation for the range}

\item{smooth.winlen}{integer, showing how many true in a row to get a true in the index,
used for cleaning fragmental: minimum allowed len for index to be positive}
}
\value{
zoo::index, monthly for now, can used by series[idx]
}
\description{
excluding values that are out of sd range (mean p/n threshold times sd)
  from the rolling window to get more stable range of changes
}
\examples{
# to filter the efa's EFA
pcg.SEE <- pe.SEE / xts::lag.xts(pe.SEE) - 1 # pe.SEE is a PE ratio xts dataframe
stable.idx.efa <- pcg.SEE$EFA \%>\% na.omit \%>\%
  GenSubidxSD(sd.threshold = 2, IF.LAG = TRUE, smooth.winlen = 3, winlen = 36)
pcg.SEE$EFA \%>\% na.omit \%>\%
  PlotFilterCompLine(stable.idx.efa, "Std Filter for EFA")

# partially modify the signal
signal[!(index(signal) \%in\% stable.idx.efa)] <- 1;

}
