% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf-metrics.R
\name{GenTrets}
\alias{GenTrets}
\title{Generate a list of total returns of the portfolio guided by the signal}
\usage{
GenTrets(weights, returns, returns.period = c("Monthly", "Daily"),
  sub.period = NULL)
}
\arguments{
\item{weights}{a list of xts dataframes, representing weights time series, generated
from GenWeights, with names properly set as signal.names}

\item{returns}{a xts dataframe with columns being assets traded in the portfolio, should
have the same number of columns as member in weights}

\item{returns.period}{string, indicating periodicty of returns, can be used to calculate daily
performance metrics if passed a daily return dataframe and monthly weights}

\item{sub.period}{vector of xts time index, can be indexed by xts objects, default NULL
can only take monthly sub.period index for now, when the returns are daily, only get
the first day of each month, raise error}
}
\value{
a list of xts series, total returns of signals, with names properly set as signal.names
}
\description{
Generate a list of total returns of the portfolio guided by the signal
}
