% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf-metrics.R
\name{GenTurnover}
\alias{GenTurnover}
\title{Generate two way turnover}
\usage{
GenTurnover(bop.weight, returns)
}
\arguments{
\item{bop.weight}{xts object of Beginning of Period Weights,
each column represents one assets}

\item{returns}{xts object of log returns, same periodicity and size
as bop.weight}
}
\value{
average monthly two-way turnover rate
}
\description{
Generate two way turnover
}
