% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet-rollmat.R
\name{GenUpperIJ}
\alias{GenUpperIJ}
\title{Generate index list for upper triangular matrix of n-order squared matrix}
\usage{
GenUpperIJ(n)
}
\description{
Generate index list for upper triangular matrix of n-order squared matrix
}
