% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mgmt.R
\name{GenVarSet}
\alias{GenVarSet}
\title{Generate a set for variables in dataframe from quotemedia}
\usage{
GenVarSet(columns, initset, asset.names)
}
\arguments{
\item{columns}{vector of string a subset of columns in dataframes
in initset from GenInitSet}

\item{initset}{list of xts dataframes from GenInitSet}

\item{asset.names}{vector of string subset of tradable assets with
full set of variables, i.e. VIX might not have volumes}
}
\value{
a list for variables, same time index as which in initset
}
\description{
Generate a set for variables in dataframe from quotemedia
}
