% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{GenVCOV}
\alias{GenVCOV}
\title{Generate vectorized var/cov dataframe from individual log returns}
\usage{
GenVCOV(ind.returns, period = c("monthly", "daily"), FUNC = NULL,
  roll.window = 21)
}
\arguments{
\item{ind.returns}{daily xts object properly named with individual
log returns, indexed using date}

\item{period}{string indicating the period in which the vcov is collected.
default: 'monthly', if 'daily', doing rolling apply, getting realized
vcov matrix from the past 21 days}

\item{FUNC}{function applied on getting vectorirzed matrix}

\item{roll.window}{integer length of window in rollapply}
}
\value{
xts object properly named with each row being vectorized
  realized variance/covariance matrix of that period
}
\description{
works the same as first getting bivariate returns dataframe,
  calculating monthly variance and covariance seperately and mergeing
}
