% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{GenVCOVName}
\alias{GenVCOVName}
\title{Generate names for vcov matrix given individual names}
\usage{
GenVCOVName(ind.names)
}
\arguments{
\item{ind.names}{vector of string, individual names}
}
\value{
vector of string, names for vectorized vcov matrix
  following row-first order and only for upper triangular
  matrix of the square vcov matrix
}
\description{
improved version using pure functional programming
  without any for loop and step-wise condition expression
}
\examples{
ind.names <- c("KW", "FO", "BY", "JZ", "DR")
GenVCOVName(ind.names)
}
