% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{GenWeights}
\alias{GenWeights}
\title{Generate a list of weight dataframes for a list of signals}
\usage{
GenWeights(signal.pack, signal.rep = c("UD", "DE"), inits = NULL,
  returns)
}
\arguments{
\item{signal.pack}{list of signals with columns properly named, the first one
will be taken as the benchmark signal, each signal should be a xts object with
one or multiple columns, signal representation should match param signal.rep}

\item{signal.rep}{string indicating signal representation, 'UD' means 2 series,
US signal and DM signal, 'DE' means 2 series, DM signal and EM signal (old DRI)}

\item{inits}{vector of numerics, representing new weights c(us.init, dm.init)}

\item{returns}{a xts dataframe with columns being assets traded in the portfolio, should
have the same number of columns as member in weights}
}
\value{
a list of weight dataframes/xts objects, with number of columns same as number of
  columns in returns (returns of assets traded in the portfolio)
}
\description{
Generate a list of weight dataframes for a list of signals
}
