% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf-metrics.R
\name{GenWsDE}
\alias{GenWsDE}
\title{Generate global weigths for US/Intl(DM/EM) using original DRI (DE rep)}
\usage{
GenWsDE(DE.signal, names, us.init = 0.6, dm.init = 6/7)
}
\arguments{
\item{DE.signal}{xts object of DRI signal (DE rep, trans to dri.dm, dri.em),
representing proportion of initial weights in global portfolio in DM (EFA)
and EM (EEM) that should be transfer to the US portfolio (original DRI signal)}

\item{names}{colnames of output, should be sth like colnames(monthly.ret)}

\item{us.init, dm.init}{numeric initial weights of us in global, dm in intl}
}
\description{
Generate global weigths for US/Intl(DM/EM) using original DRI (DE rep)
}
