% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mgmt.R
\name{GetQuoteMedia}
\alias{GetQuoteMedia}
\title{Get a raw dataframe from App.QuoteMedia.com}
\usage{
GetQuoteMedia(symbol = "SPY", start.date = as.Date("2015-01-01"),
  end.date = Sys.Date())
}
\arguments{
\item{symbol}{string asset ticker, default: 'SPY'}

\item{start.date}{date start date}

\item{end.date}{date end date}
}
\value{
raw df directly downloaded from the server, not cleaned
}
\description{
bug: start with a month later than the input date.
  fixed: allow start date to be different, use seperate conditions
  in selecting datasets (eg. in DRIVS::PlaceETFVariables)
}
