% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artillery.R
\name{GetVariable}
\alias{GetVariable}
\title{Get a DataFrame for one variable given list of symbols and dates}
\usage{
GetVariable(symbol.list = c("EQL", "GRI"),
  start.date = as.Date("2013-01-01"), end.date = as.Date("2018-07-01"),
  variable = "logreturn")
}
\arguments{
\item{symbol.list}{vector of string}

\item{start.date}{date}

\item{end.date}{date}

\item{variable}{string 'logreturn' only for now}
}
\value{
xts object with NA
}
\description{
Use DRIVS::GetXTS and period avaibility check
}
