% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artillery.R
\name{GetXTS}
\alias{GetXTS}
\title{Function Wrapper for getting an all numerical xts subject}
\usage{
GetXTS(symbol = "SPY", start.date, end.date = as.Date(Sys.time()),
  method = "QuoteMedia")
}
\arguments{
\item{symbol}{string asset ticker, default: 'SPY'}

\item{start.date}{date start date}

\item{end.date}{date end date}

\item{method}{string name of method/server for fetching data,
default: 'QuoteMedia' that calls DRIVS::GetQuoteMedia()}
}
\value{
xts object with date as index, other columns are numeric,
   NA permitted
}
\description{
This function can be used not just for ETFs
}
