% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoot.R
\name{GroupMatToShlist}
\alias{GroupMatToShlist}
\title{Trans Group Matrix to Shrinked List}
\usage{
GroupMatToShlist(mat)
}
\arguments{
\item{matrix}{showing pair-wise group/list info}
}
\value{
shrinked list of vectors
}
\description{
Trans Group Matrix to Shrinked List
}
