% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{IfHigherMedian}
\alias{IfHigherMedian}
\title{If the last observation is higher than median of overall df}
\usage{
IfHigherMedian(df)
}
\description{
If the last observation is higher than median of overall df
}
