% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-ind.R
\name{IfSpike}
\alias{IfSpike}
\title{Generate binary output for volatility spikes}
\usage{
IfSpike(df)
}
\arguments{
\item{df}{xts object of rolling vcov matrix, monthly or daily}
}
\value{
binary if the spike is positive
}
\description{
Generate binary output for volatility spikes
}
