% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{MeltSubgroups}
\alias{MeltSubgroups}
\title{Melt one ticker in subgroup list to a vector for density estimations}
\usage{
MeltSubgroups(subgroup, ticker, demean = TRUE)
}
\arguments{
\item{subgroup}{list of xts, with a specific signal condition, each xts df
contains multiple assets and just for a month (cond, signal is monthly)}

\item{ticker}{string of colname in xts df, for output}

\item{demean}{bool if demeaned in each month to see deviation}
}
\value{
a vector of merging adjusted (may be demeaned) monthly returns
}
\description{
For conditional distributions analysis in nonparametric section
}
