% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullets.R
\name{MixBmkSignal}
\alias{MixBmkSignal}
\title{Mixing a pack of signal with a single benchmark signal (i.e. DRI)}
\usage{
MixBmkSignal(pack, bmk, method = c("none", "mean", "sa", "max"))
}
\arguments{
\item{pack}{a list of signals for DM/EM both, 2 columns, properly named}

\item{bmk}{a signal for DM/EM, 2 columns, DRI in DRIVS}

\item{method}{string of method used to mix}
}
\value{
a list of signals after mixing
}
\description{
Mixing a pack of signal with a single benchmark signal (i.e. DRI)
}
