% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullets.R
\name{MixPackSignal}
\alias{MixPackSignal}
\title{Mix signals in 2 packs (DM and EM packs) and cbind}
\usage{
MixPackSignal(pack.dm, pack.em, method = c("mean", "sa", "max"))
}
\arguments{
\item{pack.dm}{list of dm signals/xts series}

\item{pack.em}{list of em signals/xts series}

\item{method}{stirng of method used}
}
\value{
a xts with 2 columns DM/EM, each column is a kind of
  mix of 2 packs
}
\description{
Mix signals in 2 packs (DM and EM packs) and cbind
}
