% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet-rollreg.R
\name{MonthlyAlpha}
\alias{MonthlyAlpha}
\title{Generate monthly alpha series from daily returns}
\usage{
MonthlyAlpha(returns, mktcol = "SPY")
}
\arguments{
\item{returns}{xts dataframe with multiple assets, daily returns}

\item{mktcol}{string of colname of market asset}
}
\value{
alpha xts dataframe with yearmon index
}
\description{
Generate monthly alpha series from daily returns
}
