% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{MonthlyBeta}
\alias{MonthlyBeta}
\title{Generate monthly beta series from daily returns}
\usage{
MonthlyBeta(returns, mktcol = "SPY")
}
\arguments{
\item{returns}{xts dataframe with multiple assets, daily returns}

\item{mktcol}{string of colname of market asset}
}
\value{
beta xts dataframe with yearmon index
}
\description{
Generate monthly beta series from daily returns
}
