% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{MonthlyRegres}
\alias{MonthlyRegres}
\title{Generate monthly last residual series from daily returns}
\usage{
MonthlyRegres(returns, mktcol = "SPY", intercept = TRUE)
}
\arguments{
\item{returns}{xts dataframe with multiple assets, daily returns}

\item{mktcol}{string of colname of market asset}

\item{intercept}{bool fit the intercept or not}
}
\value{
last residual xts dataframe with yearmon index
}
\description{
Generate monthly last residual series from daily returns
}
