% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet-rollreg.R
\name{MonthlyRSq}
\alias{MonthlyRSq}
\title{Generate monthly R-Squared series from daily returns}
\usage{
MonthlyRSq(returns, mktcol = "SPY", intercept = TRUE)
}
\arguments{
\item{returns}{xts dataframe with multiple assets, daily returns}

\item{mktcol}{string of colname of market asset}

\item{intercept}{bool fit the intercept or not}
}
\value{
rsq xts dataframe with yearmon index
}
\description{
Generate monthly R-Squared series from daily returns
}
