% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{PermMetrics}
\alias{PermMetrics}
\title{Generate metrics dataframe for permutation resampling}
\usage{
PermMetrics(signal.bmk, n, inits, asset.ret, name.bmk, signal.altbmk)
}
\arguments{
\item{signal.bmk}{single xts object, signal with DE rep}

\item{n}{int number of resampling}

\item{inits}{initial weights of the portfolio, in US, DM rep}

\item{asset.ret}{xts object of returns of assets traded in the portfolio}

\item{name.bmk}{string the name for the strategy we're interested in}

\item{signal.altbmk}{xts object alternative benchmark signal for getting relative perf}
}
\value{
metric dataframe
}
\description{
Generate metrics dataframe for permutation resampling
}
\examples{
  metrics.random <- PermMetrics(signal.hybrid, 5000, c(0, 0.34/0.49), asset.ret, 'hybrid')
}
