% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{PlotDescHist}
\alias{PlotDescHist}
\title{Plot a descriptive histogram for a series}
\usage{
PlotDescHist(vec, ticker, status, var, breaks, xlim)
}
\arguments{
\item{vec}{num series}

\item{ticker, status, var}{strings for title}

\item{breaks, xlim}{histogram args}
}
\value{
histogram plot with a proper title
}
\description{
For conditional distributions analysis in nonparametric section
}
