% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{PlotFilterCompLine}
\alias{PlotFilterCompLine}
\title{Plot a Line Plot for Std Filter (for PE ratio filter)}
\usage{
PlotFilterCompLine(series, idx, title = " ")
}
\arguments{
\item{series}{xts object, original time series, monthly}

\item{idx}{vector of yearmon objects that can used for indexing series}

\item{title}{string main title of the plot}
}
\value{
plot in the output stream
}
\description{
Plot a Line Plot for Std Filter (for PE ratio filter)
}
