% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{PlotGridMetrics}
\alias{PlotGridMetrics}
\title{Plot a Grid for Perf Metrics}
\usage{
PlotGridMetrics(metrics, file, round = 2, width = 1800, height = 480)
}
\arguments{
\item{metrics}{dataframe perf metric, gened by OutputWrapper}

\item{width, height}{int of sizes, default: 1800 x 480}

\item{path}{string of path for target image output}
}
\value{
save an image in ./images/metrics.png
}
\description{
Plot a Grid for Perf Metrics
}
