% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{PlotGridMetricsList}
\alias{PlotGridMetricsList}
\title{Plot Multiple Grids for a lit of Perf Metrics}
\usage{
PlotGridMetricsList(metrics.list, file, round = 2, width = 1400,
  height = 1600, ncol = 2)
}
\arguments{
\item{metrics.list}{a list of perf metrics dataframe
generated by DecompMetrics}

\item{file}{string of path of target file}
}
\value{
none
}
\description{
Plot Multiple Grids for a lit of Perf Metrics
}
