% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{PlotHeatMap}
\alias{PlotHeatMap}
\title{Heat Map from ggplot2 for signal pack}
\usage{
PlotHeatMap(signals, file = "./images/corr_heat.png", width = 900,
  height = 800)
}
\arguments{
\item{signals}{signal pack/list used in main scripts}
}
\description{
Heat Map from ggplot2 for signal pack
}
