% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet-rollreg.R
\name{RollingAlpha}
\alias{RollingAlpha}
\title{Rolling apply fitted alpha/intcpt series of reg: ret~ret.mkt}
\usage{
RollingAlpha(returns, winlen = 21, mktcol = "SPY")
}
\arguments{
\item{returns}{xts df of returns, one column is the mkt return
default daily, can be monthly}

\item{winlen}{length rollapply window, default 21 trading days}

\item{mktcol}{string colname of market, default 'SPY'}
}
\description{
Rolling apply fitted alpha/intcpt series of reg: ret~ret.mkt
}
