% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{RollingRegres}
\alias{RollingRegres}
\title{Rolling apply last residual of reg: ret~ret.mkt}
\usage{
RollingRegres(returns, winlen = 21, mktcol = "SPY", intercept = TRUE)
}
\arguments{
\item{returns}{xts df of returns, one column is the mkt return
default daily, can be monthly}

\item{winlen}{length rollapply window, default 21 trading days}

\item{mktcol}{string colname of market, default 'SPY'}

\item{intercept}{bool fit the intercept or not}
}
\description{
Rolling apply last residual of reg: ret~ret.mkt
}
