% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{RollingRSq}
\alias{RollingRSq}
\title{Rolling apply R-squared series of reg: ret~ret.mkt}
\usage{
RollingRSq(returns, winlen = 21, mktcol = "SPY")
}
\arguments{
\item{returns}{xts df of returns, one column is the mkt return
default daily, can be monthly}

\item{winlen}{length rollapply window, default 21 trading days}

\item{mktcol}{string colname of market, default 'SPY'}
}
\description{
Rolling apply R-squared series of reg: ret~ret.mkt
}
