% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoot.R
\name{ShlistToGroupMat}
\alias{ShlistToGroupMat}
\title{Trans Shrinked List to Group Matrix}
\usage{
ShlistToGroupMat(tmp)
}
\arguments{
\item{tmp}{list of vectors representing group info}
}
\value{
matrix showing pair-wise group/list info
}
\description{
Trans Shrinked List to Group Matrix
}
