% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet.R
\name{SignalWrapper}
\alias{SignalWrapper}
\title{Wrap up a general signal}
\usage{
SignalWrapper(FUN, params, backtest.period, name = "US", if.lag = TRUE)
}
\arguments{
\item{FUN}{function used for generating score}

\item{params}{list of params for FUN}

\item{backtest.period}{string used for indexing xts}

\item{name}{name of the signal column}

\item{if.lag}{bool if lag the signal to the next period}
}
\value{
xts signal series object
}
\description{
Wrap up a general signal
}
