% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{SplitStratName}
\alias{SplitStratName}
\title{Split the strategy name into a vector of 2}
\usage{
SplitStratName(strategy.name)
}
\arguments{
\item{strategy.name}{string of variable.METHOD,
method name (intials) should be all captitalized,
variable name can have '.' but NO uppercase letter}
}
\value{
a vector of 2 strings, variable and scoring method
}
\description{
Split the strategy name into a vector of 2
}
