% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullets.R
\name{StdBinaryPack}
\alias{StdBinaryPack}
\title{Generate a list of signals, binary pack with if higher than mean, median and quantiles}
\usage{
StdBinaryPack(monthly.var, var.name, backtest.period, IF.LAG,
  rolling.period = 36)
}
\arguments{
\item{monthly.var}{xts object, key variables to generate 0-1 signals}

\item{var.name}{string all lower case}

\item{backtest.period}{string that can used for indexing}

\item{IF.LAG}{bool}

\item{rolling.period}{int num of periods to be applied in the rolling window}
}
\value{
list of xts signal series with proper colnames
}
\description{
Generate a list of signals, binary pack with if higher than mean, median and quantiles
}
