% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullets.R
\name{StdMonthlyPack}
\alias{StdMonthlyPack}
\title{Generate a list of signals, standard pack including cont and binary signals}
\usage{
StdMonthlyPack(monthly.var, var.name, backtest.period, IF.LAG,
  rolling.period = 36)
}
\arguments{
\item{monthly.var}{xts object, key variables to generate 0-1 signals}

\item{var.name}{string all lower case}

\item{backtest.period}{string that can used for indexing}

\item{IF.LAG}{bool}

\item{rolling.period}{int num of periods to be applied in the rolling window}
}
\value{
list of xts signal series with proper colnames
}
\description{
compared to previous version in the script, change ZS and EQ to monthly calculation,
  now have higher std and higher range, more sensitive, higher corr more than 80%
}
