% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artillery.R
\name{StrToNum}
\alias{StrToNum}
\title{Convert string columns (w expressions) to numeric}
\usage{
StrToNum(raw.df, exclude = NULL)
}
\arguments{
\item{raw.df}{dataframe with string columns}

\item{exclude}{vector of strings/names of columns excluded
from the operation, default: NULL}
}
\value{
dataframe with string columns converted and evaluated
  as numeric columns
}
\description{
Deal with string columns mainly, evaluate
  expressions embedded in string columns with eval()
  and replace with numeric columns.
  Bug fixed (Aug 22): preserve dimemsion using drop = F in apply()
}
