% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tibbleXTS}
\alias{tibbleXTS}
\title{Transfer a xts object to a tibble with date (yearmon/date)}
\usage{
tibbleXTS(x, date.range = NULL)
}
\arguments{
\item{x}{xts series, default monthly}

\item{date.range}{a vector of upper and lower limit,
example for monthly data: "Jan 2010/Jan 2017"}
}
\value{
subsetted tibble dataframe
}
\description{
Transfer a xts object to a tibble with date (yearmon/date)
}
