% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{UDtoDE}
\alias{UDtoDE}
\title{Transfer US and DM ratio signals to DM and EM (original DRI)}
\usage{
UDtoDE(UD.signal, us.init = 0.6, dm.init = 6/7)
}
\description{
Transfer US and DM ratio signals to DM and EM (original DRI)
}
