% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-ind.R
\name{VecASYM}
\alias{VecASYM}
\title{Vectorize the upper tri part of asym matrix}
\usage{
VecASYM(raw.df)
}
\arguments{
\item{raw.df}{dataframe or xts object}
}
\value{
vector of numeric values, sum of pair-wise product
  output has the same order of GenVCOVName
}
\description{
Vectorize the upper tri part of asym matrix
}
