% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-ind.R
\name{VecAsymVCOV}
\alias{VecAsymVCOV}
\title{Vectorize asymmetry of covariance (pure difference)}
\usage{
VecAsymVCOV(raw.df)
}
\arguments{
\item{raw.df}{dataframe or xts object}
}
\value{
vector of vcov asymmetry based on SPY returns
}
\description{
Vectorize asymmetry of covariance (pure difference)
}
