% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-ind.R
\name{VecAsymVCOV2}
\alias{VecAsymVCOV2}
\title{Vectorize asymmetry of sum of products}
\usage{
VecAsymVCOV2(raw.df)
}
\arguments{
\item{raw.df}{dtaframe or xts object}
}
\value{
vector of redefined asymmetry based on SPY returns
}
\description{
Vectorize asymmetry of sum of products
}
