% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-ind.R
\name{VecAsymVCOV3}
\alias{VecAsymVCOV3}
\title{Vectorize asymmetry of sum of products (version 2)}
\usage{
VecAsymVCOV3(raw.df)
}
\arguments{
\item{raw.df}{dtaframe or xts object}
}
\value{
vector of redefined asymmetry^3 based on SPY returns
}
\description{
Vectorize asymmetry of sum of products (version 2)
}
