% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-ind.R
\name{VecBadVCOV}
\alias{VecBadVCOV}
\title{Vectorize covariance when the SPY is down (bad)}
\usage{
VecBadVCOV(raw.df)
}
\arguments{
\item{raw.df}{dataframe or xts object}
}
\value{
vector of bad vcov based on SPY returns
}
\description{
Vectorize covariance when the SPY is down (bad)
}
