% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-ind.R
\name{VecVCOV}
\alias{VecVCOV}
\title{Vectorize the upper tri part of vcov matrix for a dataframe}
\usage{
VecVCOV(raw.df)
}
\arguments{
\item{raw.df}{dataframe or xts object}
}
\value{
vector of numeric values in vectorized vcov matrix
  for only the upper triangular part
}
\description{
Vectorize the upper tri part of vcov matrix for a dataframe
}
