% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoot.R
\name{GenASH}
\alias{GenASH}
\title{Generate density list for groups using ASH}
\usage{
GenASH(groups, nbin = 128, range = NULL)
}
\arguments{
\item{groups}{a list of conditional variables from GenCondGroups}

\item{nbin}{number of bin hyperparameter}
}
\value{
list of density series
}
\description{
Generate density list for groups using ASH
}
