% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{GenBivRet}
\alias{GenBivRet}
\title{Generate a list for bivariate returns}
\usage{
GenBivRet(ind.returns)
}
\arguments{
\item{ind.returns}{xts object of individual returns}
}
\value{
list of xts object, each is a pair of returns
  extracted from the input ind.returns
}
\description{
Generate a list for bivariate returns
}
