% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{GenSeqValue}
\alias{GenSeqValue}
\title{Generate sequential value for the (i, j) point in (n, n) matrix}
\usage{
GenSeqValue(n, i, j)
}
\arguments{
\item{n, i, j}{int matrix dim is (n, n), point pos is (i, j)}
}
\value{
int sequential value starting from upper left, row first,
  only for upper triangular matrix of the original matrix.
  For examples, the sequential value of (2,2) in (3,3) is 4
}
\description{
Generate sequential value for the (i, j) point in (n, n) matrix
}
