% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{GenSpike}
\alias{GenSpike}
\title{Generate scaled z scores for volatility spikes}
\usage{
GenSpike(df)
}
\arguments{
\item{df}{xts object of rolling vcov matrix, monthly or daily}
}
\value{
numeric vector of (scaled) spike values
}
\description{
# some test about risk spikes
  spikes <- stats::na.omit(zoo::rollapply(monthly.vcov[, "SPY"], 36, GenSpike))
  spikes.lagged <- stats::na.omit(xts::lag.xts(spikes))["2004/2017"]
  longterm.risk <- stats::na.omit(xts::lag.xts(stats::na.omit(zoo::rollapply(monthly.vcov, 11, function(df) colMeans(df))), 2))["2004/2017"]
  last.risk <- stats::na.omit(xts::lag.xts(monthly.vcov))["2004/2017"]

  summary(lm(monthly.ret[, "SPY"] ~ last.risk[, "SPY"] + longterm.risk[, "SPY"])) # -27.01(-2.860), 34.23(2.171
  summary(lm(monthly.ret[, "EFA"] ~ last.risk[, "EFA"] + longterm.risk[, "EFA"])) # -18.64(-2.066), 33.76(2.199
  summary(lm(monthly.ret[, "EEM"] ~ last.risk[, "EEM"] + longterm.risk[, "EEM"])) # -6.73(-1.262), 3.48(2.495)
}
