% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf-metrics.R
\name{GenWsUD}
\alias{GenWsUD}
\title{Generate global weights for US/Intl(DM/EM) (UD rep)}
\usage{
GenWsUD(UD.signal, names, us.init = 0.6, dm.init = 6/7)
}
\arguments{
\item{UD.signal}{xts object with 2 columns (us.signal and dm.signal).
us.signal: xts object series of signals of US market in the overall
portfolio, ranging from 0 to 1, default: monthly periodicity.
dm.signal xts objects series of signals of DM in the international
portfolio, default: monthly periodicity, ranging from 0 to 1, calculated
in the same way as us.signal check 2 signals have the same index as monthly return}

\item{names}{colnames of output, user should double check}

\item{us.init}{numeric initial weight of us market in overall
portfolio}

\item{dm.init}{numeric initial weight of dm in international
portfolio (not overall)}
}
\value{
xts object series of weights of US, DM, EM in
  the international portfolio, using same index as input signals
}
\description{
designed for the following scenario: US/Intl(DM/EM),
  preset initial (index) weights in portfolio presented by US weight in
  overall portfolio and DM weight in international portfolio, pass in
  signal series with value ranging from 0 to 1, with the same index as
  monthly return
}
