% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{MonthlyRSq}
\alias{MonthlyRSq}
\title{Generate monthly R-Squared series from daily returns}
\usage{
MonthlyRSq(returns, mktcol = "SPY")
}
\arguments{
\item{returns}{xts dataframe with multiple assets, daily returns}

\item{mktcol}{string of colname of market asset}
}
\value{
rsq xts dataframe with yearmon index
}
\description{
Generate monthly R-Squared series from daily returns
}
