% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf-metrics.R
\name{PermMetrics}
\alias{PermMetrics}
\title{Generate metrics dataframe for permutation resampling}
\usage{
PermMetrics(signal.bmk, n = 5000, inits = c(0, 0.34/0.49), asset.ret,
  name.bmk)
}
\arguments{
\item{signal.bmk}{xts object, signal with DE rep}

\item{n}{int number of resampling (permutation)}

\item{inits}{initial weights of the portfolio}

\item{asset.ret}{xts object of returns of assets traded in the portfolio}

\item{name.bmk}{string the name for the strategy we're interested in}
}
\value{
metric dataframe
}
\description{
plotting example:
  metrics.random <- PermMetrics(signal.hybrid, 5000, inits = c(0, 0.34/0.49), asset.ret, 'hybrid')
  metric <- rownames(metrics.random)[4]
  hist(metrics.random[metric, ], breaks = 300, main = paste("Hist: ", metric, sep = "")); abline(v = metrics.random[metric, 1])
  stats::ecdf(metrics.random[metric, ])(metrics.random[metric, 1])
}
