% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{SignalWrapper}
\alias{SignalWrapper}
\title{Wrap up a general signal}
\usage{
SignalWrapper(FUN, params, backtest.period, name = "US", if.lag = TRUE)
}
\arguments{
\item{FUN}{function used for generating score}

\item{params}{list of params for FUN}

\item{backtest.period}{string used for indexing xts}

\item{name}{name of the signal column}

\item{if.lag}{bool if lag the signal to the next period}
}
\value{
xts signal series object
}
\description{
signal.cov.FS2 <- SignalWrapper(DRIVS::GenRelScore,
    list(monthly.vcov[, 'SPY & EE'], 'monthly', 36, 0,
    DRIVS::GenFeatScale, 'monthly'),
backtest.period, 'US', TRUE)

signal.cov.FS <- stats::na.omit(xts::lag.xts(
    DRIVS::GenRelScore(monthly.vcov[, 'SPY & EE'], 'monthly',
                       rolling.period = 36,
                       burn.in.period = 0,
                       scoring.method = DRIVS::GenFeatScale,
                       output.period = 'monthly')
))[backtest.period]; names(signal.cov.FS) <- "cov.FS"

all(signal.cov.FS - signal.cov.FS2 == 0) # TRUE
}
